@@ -12,17 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.0",
-    "@angular/cdk": "^12.2.12",
-    "@angular/common": "~12.2.0",
-    "@angular/compiler": "~12.2.0",
-    "@angular/core": "~12.2.0",
-    "@angular/elements": "^12.2.12",
-    "@angular/forms": "~12.2.0",
-    "@angular/material": "^12.2.12",
-    "@angular/platform-browser": "~12.2.0",
-    "@angular/platform-browser-dynamic": "~12.2.0",
-    "@angular/router": "~12.2.0",
+    "@angular/animations": "^15.2.1",
+    "@angular/cdk": "^15.2.1",
+    "@angular/common": "^15.2.1",
+    "@angular/compiler": "^15.2.1",
+    "@angular/core": "^15.2.1",
+    "@angular/elements": "^15.2.1",
+    "@angular/forms": "^15.2.1",
+    "@angular/material": "^15.2.1",
+    "@angular/platform-browser": "^15.2.1",
+    "@angular/platform-browser-dynamic": "^15.2.1",
+    "@angular/router": "^15.2.1",
     "@asymmetrik/ngx-leaflet": "^8.1.0",
     "@geoman-io/leaflet-geoman-free": "^2.11.3",
     "document-register-element": "^1.7.2",
@@ -39,9 +39,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.12",
-    "@angular/cli": "~12.2.12",
-    "@angular/compiler-cli": "~12.2.0",
+    "@angular-devkit/build-angular": "^15.2.1",
+    "@angular/cli": "^15.2.1",
+    "@angular/compiler-cli": "^15.2.1",
+    "@types/angular-material": "^1.1.73",
     "@types/google.maps": "^3.52.0",
     "@types/jasmine": "~3.8.0",
     "@types/jwt-decode": "^3.1.0",
@@ -56,6 +57,7 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.3.5"
+    "schematics-scss-migrate": "^1.3.15",
+    "typescript": "~4.9.5"
   }
 }
