@@ -29,7 +29,6 @@ import {ClipboardModule} from '@angular/cdk/clipboard';
 import {MatExpansionModule} from '@angular/material/expansion';
 
 import { MatTooltipModule } from '@angular/material/tooltip';
-import { MatChipsModule, MatChipList } from '@angular/material/chips';
 
 const modules = [
     // CommonModule,
@@ -62,7 +61,7 @@ const modules = [
     MatExpansionModule,
 
     MatTooltipModule,
-    MatChipsModule,
+
 ];
 
 @NgModule({
