@@ -4,7 +4,7 @@ import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 @Component({
   selector: 'app-alert-dialog',
   templateUrl: './alert-dialog.component.html',
-  styleUrls: ['./alert-dialog.component.css']
+  styleUrls: ['./alert-dialog.component.scss']
 })
 export class AlertDialogComponent implements OnInit {
 
