@@ -9,7 +9,7 @@ import { MqttClientService } from '../../services/mqtt-client.service';
 @Component({
   selector: 'app-navigation',
   templateUrl: './navigation.component.html',
-  styleUrls: ['./navigation.component.css']
+  styleUrls: ['./navigation.component.scss']
 })
 export class NavigationComponent  implements OnInit, OnDestroy {
 
