@@ -4,7 +4,7 @@ import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 @Component({
   selector: 'app-popup-dialog',
   templateUrl: './popup-dialog.component.html',
-  styleUrls: ['./popup-dialog.component.css']
+  styleUrls: ['./popup-dialog.component.scss']
 })
 export class PopupDialogComponent implements OnInit {
 
